pyproject.toml
[tool.poetry]
name = "retrieval_modified"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
openai = "^1.57.4"
qdrant-client = "^1.12.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
